Extract the logic to fetch the plan to a separate action
