+++ src/build.ts
@@ -23,50 +23,6 @@ async function installSnapcraft(): Promise<void> {
   await exec.exec('sudo', ['snap', 'install', 'snapcraft', '--classic'])
 }
 
-async function buildInstallCharmcraft(
-  repository: string,
-  ref: string
-): Promise<void> {
-  const workingDir = '/opt/operator-workflows/charmcraft'
-  await exec.exec('sudo', ['mkdir', workingDir, '-p'])
-  await exec.exec('sudo', ['chown', os.userInfo().username, workingDir])
-  await exec.exec('git', [
-    'clone',
-    `https://github.com/${repository}.git`,
-    '--branch',
-    ref,
-    workingDir
-  ])
-  const charmcraftSha = (
-    await exec.getExecOutput('git', ['rev-parse', 'HEAD'], { cwd: workingDir })
-  ).stdout.trim()
-  const cacheKey = `charmcraft-${charmcraftSha}`
-  const charmcraftGlob = path.join(workingDir, 'charmcraft*.snap')
-  const restored = await cache.restoreCache([charmcraftGlob], cacheKey)
-  if (!restored) {
-    await installSnapcraft()
-    core.startGroup('snapcraft pack (charmcraft)')
-    await exec.exec('snapcraft', ['--use-lxd', '--verbosity', 'trace'], {
-      cwd: workingDir
-    })
-    core.endGroup()
-  }
-  const charmcraftSnaps = await (await glob.create(charmcraftGlob)).glob()
-  if (charmcraftSnaps.length == 0) {
-    throw new Error("can't find charmcraft snap")
-  }
-  await exec.exec('sudo', [
-    'snap',
-    'install',
-    charmcraftSnaps[0],
-    '--classic',
-    '--dangerous'
-  ])
-  if (!restored) {
-    await cache.saveCache([charmcraftGlob], cacheKey)
-  }
-}
-
 interface BuildCharmParams {
   plan: BuildPlan
   charmcraftChannel: string
