+++ src/publish.ts
@@ -6,144 +6,25 @@ import * as exec from '@actions/exec'
 import * as yaml from 'js-yaml'
 
 import * as github from '@actions/github'
+import { mkdtemp, normalizePath } from './utils'
 import { Plan } from './model'
 import { DefaultArtifactClient } from '@actions/artifact'
 import fs from 'fs'
 import path from 'path'
-import os from 'os'
 
 class Publish {
   private token: string
   private charmhubToken: string
-  private octokit
-  private workflowFile: string
   private artifact
   private workingDir: string
   private resourceMapping: { [key: string]: string }
-  private workflowRunId: string
-  private identifier: string
 
   constructor() {
     this.token = core.getInput('github-token')
     this.charmhubToken = core.getInput('charmhub-token')
-    this.octokit = github.getOctokit(this.token)
-    this.workflowFile = core.getInput('workflow-file')
     this.workingDir = core.getInput('working-directory')
     this.resourceMapping = JSON.parse(core.getInput('resource-mapping'))
-    this.workflowRunId = core.getInput('workflow-run-id')
-    this.identifier = core.getInput('identifier')
     this.artifact = new DefaultArtifactClient()
-  }
-
-  mkdtemp(): string {
-    return fs.mkdtempSync(path.join(os.tmpdir(), 'artifact-'))
-  }
-
-  async findWorkflowRunId(): Promise<number> {
-    if (this.workflowRunId !== '') {
-      return parseInt(this.workflowRunId)
-    }
-    const owner = github.context.repo.owner
-    const repo = github.context.repo.repo
-    const commit = await this.octokit.rest.git.getCommit({
-      owner,
-      repo,
-      commit_sha: github.context.sha
-    })
-    const tree = commit.data.tree.sha
-    core.info(`current git tree id: ${tree}`)
-    core.info(`lookup integration test workflow: ${this.workflowFile}`)
-    const workflowResp = await this.octokit.rest.actions.getWorkflow({
-      owner,
-      repo,
-      workflow_id: this.workflowFile
-    })
-    if (workflowResp.status !== 200) {
-      throw new Error(
-        `failed to find integration workflow (${this.workflowFile}): status ${workflowResp.status}`
-      )
-    }
-    core.info(`integration test workflow id: ${workflowResp.data.id}`)
-    const runIter = this.octokit.paginate.iterator(
-      this.octokit.rest.actions.listWorkflowRuns,
-      {
-        owner,
-        repo,
-        workflow_id: workflowResp.data.id,
-        status: 'success',
-        created: `>${new Date(Date.now() - 14 * 24 * 60 * 60 * 1000).toISOString().slice(0, 10)}`
-      }
-    )
-    for await (const resp of runIter) {
-      if (resp.status !== 200) {
-        throw new Error(
-          `failed to find integration workflow run: status ${resp.status}`
-        )
-      }
-      for (const run of resp.data) {
-        if (run?.head_commit?.tree_id === tree) {
-          return run.id
-        }
-      }
-    }
-    throw new Error(
-      `Failed to find integration test workflow run on tree id (${tree}).` +
-        "Consider enabling the 'Require branches to be up to date before merging' setting to ensure that the integration tests are executed on the merged commit"
-    )
-  }
-
-  async getPlan(runId: number): Promise<Plan> {
-    core.info(`search plan artifact from workflow run: ${runId}`)
-    const artifacts = (
-      await this.octokit.paginate(
-        this.octokit.rest.actions.listWorkflowRunArtifacts,
-        {
-          owner: github.context.repo.owner,
-          repo: github.context.repo.repo,
-          run_id: runId
-        }
-      )
-    )
-      .filter(a =>
-        a.name.endsWith(
-          `${this.identifier ? '__' : ''}${this.identifier}__plan`
-        )
-      )
-      .sort((a, b) => {
-        if (a.name < b.name) {
-          return 1
-        }
-        if (a.name > b.name) {
-          return -1
-        }
-        return 0
-      })
-    for (const artifact of artifacts) {
-      const tmp = this.mkdtemp()
-      await this.artifact.downloadArtifact(artifact.id, {
-        path: tmp,
-        findBy: {
-          token: this.token,
-          repositoryOwner: github.context.repo.owner,
-          repositoryName: github.context.repo.repo,
-          workflowRunId: runId
-        }
-      })
-      const plan = JSON.parse(
-        fs.readFileSync(path.join(tmp, 'plan.json'), { encoding: 'utf-8' })
-      ) as Plan
-      if (
-        plan.working_directory === '.' ||
-        this.normalizePath(this.workingDir) ===
-          this.normalizePath(plan.working_directory) ||
-        this.normalizePath(this.workingDir).startsWith(
-          this.normalizePath(plan.working_directory) + '/'
-        )
-      ) {
-        return plan
-      }
-    }
-    throw new Error(`can't find plan artifact for workflow run ${runId}`)
   }
 
   async getCharmResources(): Promise<[string[], string[]]> {
@@ -195,15 +76,13 @@ class Publish {
     return [images, files]
   }
 
-  async getFiles(): Promise<Map<string, string>> {
+  async getFiles(plan: Plan, runId: number): Promise<Map<string, string>> {
     const [, resources] = await this.getCharmResources()
     core.info(`required resources: ${resources}`)
     const upload: Map<string, string> = new Map()
     if (resources.length === 0) {
       return upload
     }
-    const runId = await this.findWorkflowRunId()
-    const plan = await this.getPlan(runId)
     for (const build of plan.build) {
       if (build.type === 'file') {
         const resourceName = this.resourceMapping[build.name]
@@ -211,7 +90,7 @@ class Publish {
           core.info(`skip uploading file: ${build.name}`)
           continue
         }
-        const tmp = this.mkdtemp()
+        const tmp = mkdtemp()
         const artifact = await this.artifact.getArtifact(build.output, {
           findBy: {
             token: this.token,
@@ -247,15 +126,13 @@ class Publish {
     return upload
   }
 
-  async getImages(): Promise<Map<string, string>> {
+  async getImages(plan: Plan, runId: number): Promise<Map<string, string>> {
     const [resources] = await this.getCharmResources()
     core.info(`required resources: ${resources}`)
     const upload: Map<string, string> = new Map()
     if (resources.length === 0) {
       return upload
     }
-    const runId = await this.findWorkflowRunId()
-    const plan = await this.getPlan(runId)
     let dockerLogin = false
     for (const build of plan.build) {
       if (build.type === 'charm' || build.type === 'file') {
@@ -268,7 +145,7 @@ class Publish {
         core.info(`skip uploading image: ${build.name}`)
         continue
       }
-      const tmp = this.mkdtemp()
+      const tmp = mkdtemp()
       const artifact = await this.artifact.getArtifact(build.output, {
         findBy: {
           token: this.token,
@@ -342,13 +219,10 @@ class Publish {
     return upload
   }
 
-  normalizePath(p: string): string {
-    return path.normalize(p).replace(/\/+$/, '')
-  }
-
-  async getCharms(): Promise<{ name: string; dir: string; files: string[] }> {
-    const runId = await this.findWorkflowRunId()
-    const plan = await this.getPlan(runId)
+  async getCharms(
+    plan: Plan,
+    runId: number
+  ): Promise<{ name: string; dir: string; files: string[] }> {
     // FIXME: a workaround for paas app charms
     let charmDir = this.workingDir
     if (fs.existsSync(path.join(charmDir, 'charm', 'charmcraft.yaml'))) {
@@ -357,7 +231,7 @@ class Publish {
     const charms = plan.build.filter(
       b =>
         b.type === 'charm' &&
-        this.normalizePath(b.source_directory) === this.normalizePath(charmDir)
+        normalizePath(b.source_directory) === normalizePath(charmDir)
     )
     if (charms.length === 0) {
       throw new Error('no charm to upload')
@@ -368,7 +242,7 @@ class Publish {
       )
     }
     const charm = charms[0]
-    const tmp = this.mkdtemp()
+    const tmp = mkdtemp()
     const artifact = (
       await this.artifact.getArtifact(charm.output, {
         findBy: {
@@ -401,15 +275,17 @@ class Publish {
   async run() {
     try {
       core.startGroup('retrieve image info')
-      const imageResources = await this.getImages()
-      const fileResources = await this.getFiles()
+      const plan: Plan = JSON.parse(core.getInput('plan'))
+      const runId: number = +core.getInput('run-id')
+      const imageResources = await this.getImages(plan, runId)
+      const fileResources = await this.getFiles(plan, runId)
       core.endGroup()
       core.startGroup('retrieve charm info')
       const {
         name: charmName,
         dir: charmDir,
         files: charms
-      } = await this.getCharms()
+      } = await this.getCharms(plan, runId)
       core.endGroup()
       if (fileResources.size !== 0) {
         core.info(
